+++ server.js
@@ -1,15 +1,14 @@
 /* jshint node:true */
 /* global require */
 const CONFIG = require('./src/config/config.js')
-const mongoose = require('mongoose')
 const bodyParser = require('body-parser')
-const logger = require('./src/middlewares/logger.handler.js')
 const express = require('express')
-const morgan = require('morgan')
 const addRoutes = require('./src/routes/routes.js')
-const LdapAuth = require('./src/modules/authentication/LdapAuth.js')
 const sessionMiddleWare = require('./src/middlewares/session.handler.js')
-const User = require('./src/schemas/user.schema.js').User
+const addLoggerMiddleware = require('./src/middlewares/logger.handler.js')
+const ldap_initialization = require('./src/utils/ldap_initialization.js')
+const path = require('path')
+
 //app initialization
 const app = express()
 
@@ -19,63 +18,16 @@ app.use(bodyParser.json())
 app.use(bodyParser.urlencoded({ extended: false }))
 app.use(express.json())
 app.use(sessionMiddleWare)
-
-morgan.token('user', (req) => {
-  return req.user ? req.user.uid : 'anonymous'
-})
-app.use(
-  morgan(function (tokens, req, res) {
-    const log = {
-      method: tokens.method(req, res),
-      url: tokens.url(req, res),
-      status: tokens.status(req, res),
-      content_length: tokens.res(req, res, 'content-length'),
-      response_time: tokens['response-time'](req, res),
-      user: (tokens.user = req.user.uid),
-    }
-    logger.info({ ...log })
-
-    return [
-      `method:${log.method}`,
-      `url:${log.url}`,
-      `status:${log.status}`,
-      `content-lenght:${log.content_length}`,
-      `response-time:${log.response_time}ms`,
-      `user_uid:${log.user}`,
-    ].join(' ')
-  })
-)
+addLoggerMiddleware(app)
 
 //LDAP initialization
-const { usernameAttr, userOptions } = require('./src/constants/ldap_options.js')
-LdapAuth.initialize(
-  userOptions,
-  app,
-  (id) => User.findOne({ username: id }).exec(),
-  async (user) => {
-    console.log(`${user[usernameAttr]} has logged in`)
-    let foundUser = await User.findOneAndUpdate(
-      { username: user[usernameAttr] },
-      user,
-      {
-        upsert: true,
-        new: true,
-      }
-    ).exec()
-    console.log(`${foundUser.username} is retrieved from database`)
-    return foundUser
-  }
-)
+ldap_initialization(app)
 
 //add routes to application
 addRoutes(app)
 
-// use the library express-passport-ldap-mongoose
-// backward compatible mode
-/*LdapAuth.init(CONFIG.ldap.dn, CONFIG.ldap.url, app, 
-  (id) => User.findOne({ uid: id }).exec(), 
-  (user) => User.findOneAndUpdate({ uid: user.uid }, user, { upsert: true, new: true }).exec()
-)*/
+// Configure app to require modules from "src" directory
+app.set('~', path.join(__dirname, 'src'))
 
 // serve static pages
 app.use(express.static('./src/public'))
